requirements.txt
Django
djangorestframework
django-cors-headers
python-decouple
Pillow